Added color to displayable

+++ srcs/helpers/Color.hpp
@@ -13,7 +13,7 @@ struct Color {
   float b;
   float a;
 
-  Color(int r, int g, int b, int a = 1) : r(r / 255), g(g / 255), b(b / 255), a(a) {}
+  Color(int r, int g, int b, float a = 1) : r(static_cast<float>(r) / 255), g(static_cast<float>(g) / 255), b(static_cast<float>(b) / 255), a(a) {}
 
   void setColor(int r, int g, int b, int a = 1) {
     this->r = r / 255;
